+++ litmus-portal/frontend/src/models/redux/workflow.ts
@@ -44,6 +44,7 @@ export interface WorkflowData {
 export interface WorkflowManifest {
   engineYAML: string;
   manifest: string;
+  isCustomWorkflow: boolean;
 }
 
 export enum WorkflowActions {
