+++ litmus-portal/frontend/src/redux/reducers/workflow.ts
@@ -32,6 +32,7 @@ const initialState: WorkflowData = {
 const init: WorkflowManifest = {
   manifest: '',
   engineYAML: '',
+  isCustomWorkflow: false,
 };
 
 export const workflowData = createReducer<WorkflowData>(initialState, {
